Smarter completion threshold calculation (#124)
+ Calculate threshold based on the sum of the examples expected run
times* 2, because they're all completed in a single batch, so sum of the
bucket run times is the approximate lower bounds
+ Wait for at least 20sec for a job to expire, despite all other
calculations

Fixes #121

+++ spec/specwrk/web/endpoints_spec.rb
@@ -140,7 +140,7 @@ RSpec.describe Specwrk::Web::Endpoints do
 
         it { is_expected.to eq([200, {"content-type" => "application/json", "x-specwrk-status" => "1"}, [JSON.generate([{id: "a.rb:2", file_path: "a.rb", expected_run_time: 0.1}])]]) }
         it { expect { subject }.to change { pending.reload.length }.from(1).to(0) }
-        it { expect { subject }.to change { processing.reload["a.rb:2"] }.from(nil).to({completion_threshold: instance_of(Integer), expected_run_time: 0.1, file_path: "a.rb", id: "a.rb:2"}) }
+        it { expect { subject }.to change { processing.reload["a.rb:2"] }.from(nil).to({completion_threshold: instance_of(Float), expected_run_time: 0.1, file_path: "a.rb", id: "a.rb:2"}) }
       end
 
       context "no items in any queue" do
@@ -218,7 +218,7 @@ RSpec.describe Specwrk::Web::Endpoints do
 
         it { is_expected.to eq([200, {"content-type" => "application/json", "x-specwrk-status" => "0"}, [JSON.generate([{id: "a.rb:2", file_path: "a.rb", expected_run_time: 0.1}])]]) }
         it { expect { subject }.to change { pending.reload.length }.from(1).to(0) }
-        it { expect { subject }.to change { processing.reload["a.rb:2"] }.from(nil).to({completion_threshold: instance_of(Integer), expected_run_time: 0.1, file_path: "a.rb", id: "a.rb:2"}) }
+        it { expect { subject }.to change { processing.reload["a.rb:2"] }.from(nil).to({completion_threshold: instance_of(Float), expected_run_time: 0.1, file_path: "a.rb", id: "a.rb:2"}) }
       end
 
       context "no items in the processing queue, but completed queue has items" do
